# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ugSOT Newsletter Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AdminSession {
  email: string;
  loggedIn: boolean;
}

export interface Employee {
  id: number;
  employeeName: string;
  employeeEmail: string;
  createdAt: string;
}

export interface EmployeeListResponse {
  employees: Employee[];
  total: number;
  page: number;
  pageSize: number;
}

export interface EmployeeUploadResult {
  added: number;
  skipped: number;
  invalid: number;
  errors: string[];
}

export interface Newsletter {
  id: number;
  title: string;
  topic: string;
  /** @nullable */
  description?: string | null;
  pdfUrl: string;
  uploadedAt: string;
  /** @nullable */
  totalSent?: number | null;
  /** @nullable */
  totalFailed?: number | null;
}

export interface NewsletterListResponse {
  newsletters: Newsletter[];
  total: number;
  page: number;
  pageSize: number;
}

export interface SendResult {
  sent: number;
  failed: number;
  total: number;
}

export type EmailLogDeliveryStatus =
  (typeof EmailLogDeliveryStatus)[keyof typeof EmailLogDeliveryStatus];

export const EmailLogDeliveryStatus = {
  sent: "sent",
  failed: "failed",
  pending: "pending",
} as const;

export interface EmailLog {
  id: number;
  employeeEmail: string;
  newsletterId: number;
  /** @nullable */
  newsletterTitle?: string | null;
  deliveryStatus: EmailLogDeliveryStatus;
  sentAt: string;
  /** @nullable */
  errorMessage?: string | null;
}

export interface EmailLogListResponse {
  logs: EmailLog[];
  total: number;
  page: number;
  pageSize: number;
}

export interface DashboardStats {
  totalEmployees: number;
  totalNewsletters: number;
  totalEmailsSent: number;
  totalEmailsFailed: number;
  deliveryRate: number;
  recentNewsletters: Newsletter[];
}

export type ListEmployeesParams = {
  search?: string;
  page?: number;
  pageSize?: number;
};

export type ListNewslettersParams = {
  page?: number;
  pageSize?: number;
};

export type ListEmailLogsParams = {
  newsletterId?: number;
  status?: ListEmailLogsStatus;
  page?: number;
  pageSize?: number;
};

export type ListEmailLogsStatus =
  (typeof ListEmailLogsStatus)[keyof typeof ListEmailLogsStatus];

export const ListEmailLogsStatus = {
  sent: "sent",
  failed: "failed",
  pending: "pending",
} as const;