# lib/api-zod/src/generated/types/emailLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ugSOT Newsletter Management API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailLogDeliveryStatus } from "./emailLogDeliveryStatus";

export interface EmailLog {
  id: number;
  employeeEmail: string;
  newsletterId: number;
  /** @nullable */
  newsletterTitle?: string | null;
  deliveryStatus: EmailLogDeliveryStatus;
  sentAt: Date;
  /** @nullable */
  errorMessage?: string | null;
}