# lib/api-zod/src/generated/types/employeeListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ugSOT Newsletter Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Employee } from "./employee";

export interface EmployeeListResponse {
  employees: Employee[];
  total: number;
  page: number;
  pageSize: number;
}