# lib/api-zod/src/generated/types/newsletter.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ugSOT Newsletter Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Newsletter {
  id: number;
  title: string;
  topic: string;
  /** @nullable */
  description?: string | null;
  pdfUrl: string;
  uploadedAt: Date;
  /** @nullable */
  totalSent?: number | null;
  /** @nullable */
  totalFailed?: number | null;
}