# lib/api-zod/src/generated/types/newsletterListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ugSOT Newsletter Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Newsletter } from "./newsletter";

export interface NewsletterListResponse {
  newsletters: Newsletter[];
  total: number;
  page: number;
  pageSize: number;
}